+++ src/Sophia_Script_for_Windows_11_PowerShell_7/Module/Sophia.psm1
@@ -2,8 +2,8 @@
 	.SYNOPSIS
 	Sophia Script is a PowerShell module for Windows 10 & Windows 11 fine-tuning and automating the routine tasks
 
-	Version: v6.5.0
-	Date: 27.05.2023
+	Version: v6.5.1
+	Date: 03.06.2023
 
 	Copyright (c) 2014—2023 farag
 	Copyright (c) 2019—2023 farag & Inestic
@@ -3202,6 +3202,25 @@ function TaskViewButton
 	.NOTES
 	Current user
 #>
+<#
+	.SYNOPSIS
+	The Chat icon (Microsoft Teams) on the taskbar
+
+	.PARAMETER Hide
+	Hide the Chat icon (Microsoft Teams) on the taskbar and prevent Microsoft Teams from installing for new users
+
+	.PARAMETER Show
+	Show the Chat icon (Microsoft Teams) on the taskbar and remove block for Microsoft Teams from installing for new users
+
+	.EXAMPLE
+	TaskbarChat -Hide
+
+	.EXAMPLE
+	TaskbarChat -Show
+
+	.NOTES
+	Current user
+#>
 function TaskbarChat
 {
 	param
@@ -3221,17 +3240,39 @@ function TaskbarChat
 		$Show
 	)
 
+	Clear-Variable -Name Task -ErrorAction Ignore
+
 	switch ($PSCmdlet.ParameterSetName)
 	{
 		"Hide"
 		{
 			New-ItemProperty -Path HKCU:\Software\Microsoft\Windows\CurrentVersion\Explorer\Advanced -Name TaskbarMn -PropertyType DWord -Value 0 -Force
+			# Save string to run it as "NT SERVICE\TrustedInstaller"
+			# Prevent Microsoft Teams from installing for new users
+			$Task = "New-ItemProperty -Path HKLM:\SOFTWARE\Microsoft\Windows\CurrentVersion\Communications -Name ConfigureChatAutoInstall -Value 0 -Type Dword -Force"
 		}
 		"Show"
 		{
 			New-ItemProperty -Path HKCU:\Software\Microsoft\Windows\CurrentVersion\Explorer\Advanced -Name TaskbarMn -PropertyType DWord -Value 1 -Force
+			# Save string to run it as "NT SERVICE\TrustedInstaller"
+			# Remove block for Microsoft Teams from installing for new users
+			$Task = "Remove-ItemProperty -Path HKLM:\SOFTWARE\Microsoft\Windows\CurrentVersion\Communications -Name ConfigureChatAutoInstall -Value 1 -Type Dword -Force"
 		}
 	}
+
+	# Create a Scheduled Task to run it as "NT SERVICE\TrustedInstaller"
+	$Parameters = @{
+		TaskName = "BlockTeamsInstallation"
+		Action   = New-ScheduledTaskAction -Execute powershell.exe -Argument "-WindowStyle Hidden -Command $Task"
+	}
+	Register-ScheduledTask @Parameters -Force
+
+	$ScheduleService = New-Object -ComObject Schedule.Service
+	$ScheduleService.Connect()
+	$ScheduleService.GetFolder("\").GetTask("BlockTeamsInstallation").RunEx($null, 0, 0, "NT SERVICE\TrustedInstaller")
+
+	# Remove temporary task
+	Unregister-ScheduledTask -TaskName BlockTeamsInstallation -Confirm:$false
 }
 
 <#
@@ -9451,9 +9492,12 @@ function Export-Associations
 					$PartProgramPath = (Get-ItemPropertyValue -Path "HKCU:\SOFTWARE\Classes\$($_.ProgId)\Shell\Open\Command" -Name "(default)").Trim()
 					$Program = $PartProgramPath.Substring(0, ($PartProgramPath.IndexOf(".exe") + 4)).Trim('"')
 
-					if (Test-Path -Path $Program)
+					if ($Program)
 					{
-						$ProgramPath = $PartProgramPath
+						if (Test-Path -Path $([System.Environment]::ExpandEnvironmentVariables($Program)))
+						{
+							$ProgramPath = $PartProgramPath
+						}
 					}
 				}
 				elseif ([Microsoft.Win32.Registry]::GetValue("HKEY_LOCAL_MACHINE\SOFTWARE\Classes\$($_.ProgId)\Shell\Open\Command", "", $false))
@@ -9461,9 +9505,12 @@ function Export-Associations
 					$PartProgramPath = (Get-ItemPropertyValue -Path "HKLM:\SOFTWARE\Classes\$($_.ProgId)\Shell\Open\Command" -Name "(default)").Trim()
 					$Program = $PartProgramPath.Substring(0, ($PartProgramPath.IndexOf(".exe") + 4)).Trim('"')
 
-					if (Test-Path -Path $Program)
+					if ($Program)
 					{
-						$ProgramPath = $PartProgramPath
+						if (Test-Path -Path $([System.Environment]::ExpandEnvironmentVariables($Program)))
+						{
+							$ProgramPath = $PartProgramPath
+						}
 					}
 				}
 
@@ -9487,9 +9534,12 @@ function Export-Associations
 						}
 					}
 
-					if (Test-Path -Path $IconPath)
+					if ($IconPath)
 					{
-						$Icon = $IconPartPath
+						if (Test-Path -Path $([System.Environment]::ExpandEnvironmentVariables($IconPath)))
+						{
+							$Icon = $IconPartPath
+						}
 					}
 				}
 				elseif ([Microsoft.Win32.Registry]::GetValue("HKEY_LOCAL_MACHINE\SOFTWARE\Classes\$($_.ProgId)\DefaultIcon", "", $false))
@@ -9511,9 +9561,12 @@ function Export-Associations
 						}
 					}
 
-					if (Test-Path -Path $IconPath)
+					if ($IconPath)
 					{
-						$Icon = $IconPartPath
+						if (Test-Path -Path $([System.Environment]::ExpandEnvironmentVariables($IconPath)))
+						{
+							$Icon = $IconPartPath
+						}
 					}
 				}
 				elseif ([Microsoft.Win32.Registry]::GetValue("HKEY_CURRENT_USER\Software\Classes\$($_.ProgId)\shell\open\command", "", $false))
@@ -9521,9 +9574,12 @@ function Export-Associations
 					$IconPartPath = (Get-ItemPropertyValue -Path "HKCU:\Software\Classes\$($_.ProgId)\shell\open\command" -Name "(default)").Trim()
 					$IconPath = $IconPartPath.Substring(0, $IconPartPath.IndexOf(".exe") + 4).Trim('"')
 
-					if (Test-Path -Path $IconPath)
+					if ($IconPath)
 					{
-						$Icon = "$IconPath,0"
+						if (Test-Path -Path $([System.Environment]::ExpandEnvironmentVariables($IconPath)))
+						{
+							$Icon = "$IconPath,0"
+						}
 					}
 				}
 				elseif ([Microsoft.Win32.Registry]::GetValue("HKEY_LOCAL_MACHINE\SOFTWARE\Classes\$($_.ProgId)\Shell\Open\Command", "", $false))
@@ -9531,9 +9587,12 @@ function Export-Associations
 					$IconPartPath = (Get-ItemPropertyValue -Path "HKLM:\SOFTWARE\Classes\$($_.ProgId)\Shell\Open\Command" -Name "(default)").Trim()
 					$IconPath = $IconPartPath.Substring(0, $IconPartPath.IndexOf(".exe") + 4)
 
-					if (Test-Path -Path $IconPath)
+					if ($IconPath)
 					{
-						$Icon = "$IconPath,0"
+						if (Test-Path -Path $([System.Environment]::ExpandEnvironmentVariables($IconPath)))
+						{
+							$Icon = "$IconPath,0"
+						}
 					}
 				}
 			}
@@ -9559,6 +9618,8 @@ function Export-Associations
 "@ | ConvertFrom-JSON
 		$AllJSON += $JSON
 	}
+
+	Clear-Variable -Name ProgramPath, Icon -ErrorAction Ignore
 
 	$AllJSON | ConvertTo-Json | Set-Content -Path "$PSScriptRoot\..\Application_Associations.json" -Force -Encoding utf8